requirements.txt
ccxt>=4.0.0
vectorbt>=0.25.0
backtrader>=1.9.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
python-twitter>=3.5
praw>=7.7.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0
